fragmentation of event_routes.py
- added new blueprints (paths to endpoints are still the same)
- created 4 new files
- added description of get_event() function
